refactor: useDropdown 훅 작성
